Move ignored cve features to own components. Show only for super admins
