+++ frontend/src/guidance/AdditionalFindings.js
@@ -20,17 +20,17 @@ import {
   ModalCloseButton,
   Link,
   SimpleGrid,
-  useToast,
 } from '@chakra-ui/react'
 import { ExternalLinkIcon } from '@chakra-ui/icons'
 import { Trans, t } from '@lingui/macro'
 import { any, bool, string } from 'prop-types'
 import { useLingui } from '@lingui/react'
-import { useMutation, useQuery } from '@apollo/client'
+import { useQuery } from '@apollo/client'
 import { GUIDANCE_ADDITIONAL_FINDINGS } from '../graphql/queries'
 import { LoadingMessage } from '../components/LoadingMessage'
 import { ErrorFallbackMessage } from '../components/ErrorFallbackMessage'
-import { IGNORE_CVE, UNIGNORE_CVE } from '../graphql/mutations'
+import CveIgnorer from './CveIgnorer'
+import IgnoredCves from './IgnoredCves'
 
 export function AdditionalFindings({ domain, cveDetected }) {
   const { i18n } = useLingui()
@@ -39,111 +39,11 @@ export function AdditionalFindings({ domain, cveDetected }) {
   const [activeCve, setActiveCve] = useState('')
   const { isOpen, onOpen, onClose } = useDisclosure()
   const { isOpen: cveIsOpen, onOpen: cveOnOpen, onClose: cveOnClose } = useDisclosure()
-  const [showConfirm, setShowConfirm] = useState(false)
-  const toast = useToast()
 
   const formatTimestamp = (datetime) => new Date(datetime).toLocaleDateString()
 
   const { data, loading, error } = useQuery(GUIDANCE_ADDITIONAL_FINDINGS, {
     variables: { domain },
-  })
-
-  const [ignoreCve] = useMutation(IGNORE_CVE, {
-    refetchQueries: ['GuidanceAdditionalFindings'],
-    onError(error) {
-      toast({
-        title: i18n._(t`An error occurred.`),
-        description: error.message,
-        status: 'error',
-        duration: 9000,
-        isClosable: true,
-        position: 'top-left',
-      })
-    },
-    onCompleted({ ignoreCve }) {
-      if (ignoreCve.result.__typename === 'Domain') {
-        toast({
-          title: i18n._(t`CVE ignored`),
-          description: i18n._(
-            t`Successfully ignored CVE for domain ${ignoreCve.result.domain}. New ignored CVEs: "${
-              ignoreCve.result.ignoredCves && JSON.stringify(ignoreCve.result.ignoredCves)
-            }".`,
-          ),
-          status: 'success',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-      } else if (ignoreCve.result.__typename === 'DomainError') {
-        toast({
-          title: i18n._(t`Unable to ignore CVE.`),
-          description: ignoreCve.result.description,
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-      } else {
-        toast({
-          title: i18n._(t`Incorrect send method received.`),
-          description: i18n._(t`Incorrect ignoreCve.result typename.`),
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-        console.log('Incorrect ignoreCve.result typename.')
-      }
-    },
-  })
-
-  const [unignoreCve] = useMutation(UNIGNORE_CVE, {
-    refetchQueries: ['GuidanceAdditionalFindings'],
-    onError(error) {
-      toast({
-        title: i18n._(t`An error occurred.`),
-        description: error.message,
-        status: 'error',
-        duration: 9000,
-        isClosable: true,
-        position: 'top-left',
-      })
-    },
-    onCompleted({ unignoreCve }) {
-      if (unignoreCve.result.__typename === 'Domain') {
-        toast({
-          title: i18n._(t`Stopped ignoring CVE`),
-          description: i18n._(
-            t`Successfully stopped ignoring CVE for domain "${unignoreCve.result.domain}". New ignored CVEs: "${
-              unignoreCve.result.ignoredCves && JSON.stringify(unignoreCve.result.ignoredCves)
-            }".`,
-          ),
-          status: 'success',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-      } else if (unignoreCve.result.__typename === 'DomainError') {
-        toast({
-          title: i18n._(t`Unable to stop ignoring CVE.`),
-          description: unignoreCve.result.description,
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-      } else {
-        toast({
-          title: i18n._(t`Incorrect send method received.`),
-          description: i18n._(t`Incorrect unignoreCve.result typename.`),
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-        console.log('Incorrect unignoreCve.result typename.')
-      }
-    },
   })
 
   if (loading) {
@@ -180,14 +80,12 @@ export function AdditionalFindings({ domain, cveDetected }) {
     ({ webComponentCategory }) => !['Framework', 'DDOS Protection', 'CDN'].includes(webComponentCategory),
   )
 
-  const handleCveOnClose = () => {
-    setShowConfirm(false)
-    setActiveCve('')
-    cveOnClose()
-  }
-
-  const isCveIgnored = (cve) => {
-    return ignoredCves?.includes(cve)
+  const setActiveCveHandler = (cve) => {
+    setActiveCve({
+      cve,
+      affectedWebComps: webComponents.filter(({ webComponentCves }) => webComponentCves.some((x) => x.cve === cve)),
+    })
+    cveOnOpen()
   }
 
   return (
@@ -244,13 +142,7 @@ export function AdditionalFindings({ domain, cveDetected }) {
                                 size="sm"
                                 _hover={{ bg: cveSeverityOnHover[severity] }}
                                 onClick={() => {
-                                  setActiveCve({
-                                    cve,
-                                    affectedWebComps: webComponents.filter(({ webComponentCves }) =>
-                                      webComponentCves.some((x) => x.cve === cve),
-                                    ),
-                                  })
-                                  cveOnOpen()
+                                  setActiveCveHandler(cve)
                                 }}
                               >
                                 {cve}
@@ -267,49 +159,13 @@ export function AdditionalFindings({ domain, cveDetected }) {
                   <Trans>No Top 25 Vulnerabilites Detected</Trans>
                 </Text>
               )}
-              <Divider borderBottomColor="gray.900" />
-              <Flex flexDirection="column">
-                <Box px="2" mb="2">
-                  <Text fontWeight="bold">
-                    <Trans>Ignored CVEs:</Trans>
-                  </Text>
-                  {!ignoredCves || ignoredCves.length === 0 ? (
-                    <Text>
-                      <Trans>None</Trans>
-                    </Text>
-                  ) : (
-                    <SimpleGrid columns={8}>
-                      {ignoredCves &&
-                        ignoredCves.map((cve) => {
-                          return (
-                            <Button
-                              key={`ignored-${cve}`}
-                              borderRadius="full"
-                              m="1"
-                              borderColor="black"
-                              borderWidth="1px"
-                              bg="gray.100"
-                              fontWeight="normal"
-                              size="sm"
-                              _hover={{ bg: 'gray.200' }}
-                              onClick={() => {
-                                setActiveCve({
-                                  cve,
-                                  affectedWebComps: webComponents.filter(({ webComponentCves }) =>
-                                    webComponentCves.some((x) => x.cve === cve),
-                                  ),
-                                })
-                                cveOnOpen()
-                              }}
-                            >
-                              {cve}
-                            </Button>
-                          )
-                        })}
-                    </SimpleGrid>
-                  )}
-                </Box>
-              </Flex>
+              <IgnoredCves
+                ignoredCves={ignoredCves}
+                setActiveCve={setActiveCve}
+                webComponents={webComponents}
+                cveOnOpen={cveOnOpen}
+                setActiveCveHandler={setActiveCveHandler}
+              />
             </AccordionPanel>
           </AccordionItem>
 
@@ -537,7 +393,7 @@ export function AdditionalFindings({ domain, cveDetected }) {
         </Accordion>
       </Box>
 
-      <Modal isOpen={cveIsOpen} onClose={handleCveOnClose}>
+      <Modal isOpen={cveIsOpen} onClose={cveOnClose}>
         <ModalOverlay />
         <ModalContent>
           <ModalHeader>{activeCve?.cve}</ModalHeader>
@@ -549,62 +405,7 @@ export function AdditionalFindings({ domain, cveDetected }) {
                 {webComponentName} {webComponentCategory} {webComponentVersion}
               </Text>
             ))}
-            <Divider borderBottomColor="gray.900" />
-            <Box>
-              <Button
-                variant="primary"
-                mr="4"
-                type="submit"
-                isLoading={false}
-                px={8}
-                onClick={() => setShowConfirm(true)}
-              >
-                {isCveIgnored(activeCve.cve) ? <Trans>Stop Ignoring CVE</Trans> : <Trans>Ignore CVE</Trans>}
-              </Button>
-              {showConfirm && (
-                <Box mt="4">
-                  <Text mb="4">
-                    {isCveIgnored(activeCve.cve) ? (
-                      <Trans>Are you sure you want to stop ignoring this CVE?</Trans>
-                    ) : (
-                      <Trans>Are you sure you want to ignore this CVE?</Trans>
-                    )}
-                  </Text>
-
-                  <Button variant="primaryOutline" mr="4" onClick={() => setShowConfirm(false)}>
-                    <Trans>Back</Trans>
-                  </Button>
-
-                  <Button
-                    variant="primary"
-                    mr="4"
-                    type="submit"
-                    isLoading={false}
-                    px={8}
-                    onClick={async () => {
-                      if (isCveIgnored(activeCve.cve)) {
-                        await unignoreCve({
-                          variables: {
-                            domainId,
-                            ignoredCve: activeCve.cve,
-                          },
-                        })
-                      } else {
-                        await ignoreCve({
-                          variables: {
-                            domainId,
-                            ignoredCve: activeCve.cve,
-                          },
-                        })
-                      }
-                      setShowConfirm(false)
-                    }}
-                  >
-                    <Trans>Confirm</Trans>
-                  </Button>
-                </Box>
-              )}
-            </Box>
+            <CveIgnorer cve={activeCve?.cve} isCveIgnored={ignoredCves?.includes(activeCve?.cve)} domainId={domainId} />
           </ModalBody>
           <ModalFooter>
             <Link color="blue.500" href={`https://www.cve.org/CVERecord?id=${activeCve?.cve}`} isExternal>
